Structure of folders defined: now the components contains its own folders, including the index file and the css to itself
