+++ frontend/src/App.css
@@ -1,7 +1,12 @@
 /*Here goes the css of the application*/
 @import url('https://fonts.googleapis.com/css?family=Roboto&display=swap');
+
+
 * {
-box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+  outline: none;
+  box-sizing: border-box;
 }
 
 html, body{
@@ -29,6 +34,7 @@ h1.projectTitle {
   font-weight: normal;
   color: #888;
   display: flex;
+  flex-direction: column;
   align-items: center;
   justify-content: center;
 }
@@ -43,8 +49,7 @@ h1.projectTitle p span {
 }
 
 h1.projectTitle img {
-  height: 40px;
-  margin: 10px;
+  height: 80px;
 }
 
 a{
@@ -55,4 +60,39 @@ a{
 
 a:hover{
   color: #2db0d8;
+}
+
+label {
+  color: #888;
+}
+
+input {
+  width: 100%;
+  padding: 10px;
+  border: none;
+  border: solid 1px #ccc;
+  border-radius: 5px;
+  margin: 5px 0;
+  font-size: 15px;
+  color: #888;
+  font-size: 18px;
+  box-shadow: 0 0 15px rgba(0, 0, 0, 0.15);
+}
+
+button {
+  width: 100%;
+  padding: 10px;
+  border: none !important;
+  margin: 5px 0;
+  color: white;
+  cursor: pointer;
+  background-color: #2f94e0;
+  border-radius: 5px;
+  font-size: 15px;
+  font-size: 18px;
+  transition: filter ease-in-out 0.2s;
+}
+
+button:hover{
+  filter: brightness(130%)
 }
