--- frontend/src/css/Login.css
@@ -1,49 +1,0 @@
-div.App {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgb(236, 241, 253);
-    padding: 5px;
-}
-
-div.LoginForm {
-    width: 100%;
-    max-width: 420px;
-    background-color: white;
-    border-radius: 5px;
-    box-shadow: #bbb 0 0 35px;
-    padding: 20px;
-}
-
-div.LoginForm label {
-    color: #888;
-}
-
-div.LoginForm input {
-    width: 100%;
-    padding: 10px;
-    border: none;
-    border: solid 2px #ccc;
-    border-radius: 5px;
-    margin: 5px 0;
-    font-size: 15px;
-    color: #888;
-}
-
-div.LoginForm button {
-    width: 100%;
-    padding: 10px;
-    border: none;
-    margin: 5px 0;
-    color: white;
-    cursor: pointer;
-    background-color: #2f94e0;
-    border-radius: 5px;
-    font-size: 15px;
-    transition: background-color ease-in-out 0.2s;
-}
-
-div.LoginForm button:hover{
-    background-color: #2db0d8;
-}
